Print the Output of Intermediate Steps of a Call

boom() will explode a call into the output of its parts. rig() a function to make all the calls of its body boom().

boom(expr, clock = getOption("boom.clock"), print = getOption("boom.print"))

rig(fun, clock = getOption("boom.clock"), print = getOption("boom.print"))

Arguments

expr
call to explode

clock
whether to time intermediate steps, FALSE by default unless you set options(boom.clock = TRUE). The execution time of a step doesn't include the execution time of its previously printed sub-steps.

print
A function, a formula or a list of functions or formulas.

fun
function ro rig()

Details

By default, unless options(boom.clock = print) is set to a custom value, the output of every step is printed in a standard way.

If you provide another function such as str, the console output of str will be printed. Use invisible to display nothing, another useful alternative would be dplyr::glimpse.

rlang's formula notation is supported, so for instance you can type: print = ~ dplyr::glimpse(., width = 50).

Sometimes you might want to print a specific type of object in a custom way, in this case you can provide a named list, if you provide an unnamed element it will be used as the default, and named elements will define how objects of the given S3 class are printed. For instance print = list(str, data.frame = tibble::as_tibble)

Modifying the options is especially useful if you want to use the addin with custom behavior.

Examples

# explode a simple call
boom(subset(head(mtcars, 2), qsec > 17))

#> head(mtcars, 2)
#>               mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4      21   6  160 110  3.9 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag  21   6  160 110  3.9 2.875 17.02  0  1    4    4
#> qsec > 17
#> [1] FALSE  TRUE
#> subset(head(mtcars, 2), qsec > 17)
#>               mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4 Wag  21   6  160 110  3.9 2.875 17.02  0  1    4    4

# clock calls and customize how to print output
boom(subset(head(mtcars, 2), qsec > 17), clock = TRUE, print = str)

#> head(mtcars, 2)
#> time: 0 us
#> 'data.frame':	2 obs. of  11 variables:
#>  $ mpg : num  21 21
#>  $ cyl : num  6 6
#>  $ disp: num  160 160
#>  $ hp  : num  110 110
#>  $ drat: num  3.9 3.9
#>  $ wt  : num  2.62 2.88
#>  $ qsec: num  16.5 17
#>  $ vs  : num  0 0
#>  $ am  : num  1 1
#>  $ gear: num  4 4
#>  $ carb: num  4 4
#> qsec > 17
#> time: 0 us
#>  logi [1:2] FALSE TRUE
#> subset(head(mtcars, 2), qsec > 17)
#> time: 0.727 ms
#> 'data.frame':	1 obs. of  11 variables:
#>  $ mpg : num 21
#>  $ cyl : num 6
#>  $ disp: num 160
#>  $ hp  : num 110
#>  $ drat: num 3.9
#>  $ wt  : num 2.88
#>  $ qsec: num 17
#>  $ vs  : num 0
#>  $ am  : num 1
#>  $ gear: num 4
#>  $ carb: num 4

# print str only for data frames
boom(subset(head(mtcars, 2), qsec > 17), print = list(data.frame = str))

#> head(mtcars, 2)
#> 'data.frame':	2 obs. of  11 variables:
#>  $ mpg : num  21 21
#>  $ cyl : num  6 6
#>  $ disp: num  160 160
#>  $ hp  : num  110 110
#>  $ drat: num  3.9 3.9
#>  $ wt  : num  2.62 2.88
#>  $ qsec: num  16.5 17
#>  $ vs  : num  0 0
#>  $ am  : num  1 1
#>  $ gear: num  4 4
#>  $ carb: num  4 4
#> qsec > 17
#> [1] FALSE  TRUE
#> subset(head(mtcars, 2), qsec > 17)
#> 'data.frame':	1 obs. of  11 variables:
#>  $ mpg : num 21
#>  $ cyl : num 6
#>  $ disp: num 160
#>  $ hp  : num 110
#>  $ drat: num 3.9
#>  $ wt  : num 2.88
#>  $ qsec: num 17
#>  $ vs  : num 0
#>  $ am  : num 1
#>  $ gear: num 4
#>  $ carb: num 4

# rig a function
rig(ave)(warpbreaks$breaks, warpbreaks$wool)

#> missing(...)
#> [1] FALSE
#> interaction(...)
#>  [1] A A A A A A A A A A A A A A A A A A A A A A A A A A A B B B B B B B B B B B
#> [39] B B B B B B B B B B B B B B B B
#> Levels: A B
#> g <- interaction(...)
#>  [1] A A A A A A A A A A A A A A A A A A A A A A A A A A A B B B B B B B B B B B
#> [39] B B B B B B B B B B B B B B B B
#> Levels: A B
#> split(x, g)
#> $A
#>  [1] 26 30 54 25 70 52 51 26 67 18 21 29 17 12 18 35 30 36 36 21 24 18 10 43 28
#> [26] 15 26
#> 
#> $B
#>  [1] 27 14 29 19 29 31 41 20 44 42 26 19 16 39 28 21 39 29 20 21 24 17 13 15 15
#> [26] 16 28
#> 
#> lapply(split(x, g), FUN)
#> $A
#> [1] 31.03704
#> 
#> $B
#> [1] 25.25926
#> 
#> split(x, g) <- lapply(split(x, g), FUN)
#> $A
#> [1] 31.03704
#> 
#> $B
#> [1] 25.25926
#> 
#> {
#>     g <- interaction(...)
#>     split(x, g) <- lapply(split(x, g), FUN)
#> }
#> $A
#> [1] 31.03704
#> 
#> $B
#> [1] 25.25926
#> 
#> if (missing(...)) x[] <- FUN(x) else {
#>     g <- interaction(...)
#>     split(x, g) <- lapply(split(x, g), FUN)
#> }
#> $A
#> [1] 31.03704
#> 
#> $B
#> [1] 25.25926
#> 
#> {
#>     if (missing(...)) 
#>         x[] <- FUN(x)
#>     else {
#>         g <- interaction(...)
#>         split(x, g) <- lapply(split(x, g), FUN)
#>     }
#>     x
#> }
#>  [1] 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704
#>  [9] 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704
#> [17] 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704
#> [25] 31.03704 31.03704 31.03704 25.25926 25.25926 25.25926 25.25926 25.25926
#> [33] 25.25926 25.25926 25.25926 25.25926 25.25926 25.25926 25.25926 25.25926
#> [41] 25.25926 25.25926 25.25926 25.25926 25.25926 25.25926 25.25926 25.25926
#> [49] 25.25926 25.25926 25.25926 25.25926 25.25926 25.25926

#>  [1] 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704
#>  [9] 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704
#> [17] 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704 31.03704
#> [25] 31.03704 31.03704 31.03704 25.25926 25.25926 25.25926 25.25926 25.25926
#> [33] 25.25926 25.25926 25.25926 25.25926 25.25926 25.25926 25.25926 25.25926
#> [41] 25.25926 25.25926 25.25926 25.25926 25.25926 25.25926 25.25926 25.25926
#> [49] 25.25926 25.25926 25.25926 25.25926 25.25926 25.25926